# Performance improvement in make_structured_array