--- a/tonic/datasets/davisdataset.py
+++ b/tonic/datasets/davisdataset.py
@@ -4,6 +4,7 @@ from numpy.lib import recfunctions
 from importRosbag.importRosbag import importRosbag
 from tonic.dataset import Dataset
 from tonic.download_utils import check_integrity, download_url
+from tonic.io import make_structured_array
 
 
 class DAVISDATA(Dataset):
@@ -96,10 +97,14 @@ class DAVISDATA(Dataset):
         events = topics["/dvs/events"]
         events["ts"] -= events["ts"][0]
         events["ts"] *= 1e6
-        events = np.column_stack(
-            (events["ts"], events["x"], events["y"], events["pol"])
+        events = make_structured_array(
+            events["ts"], 
+            events["x"], 
+            events["y"], 
+            events["pol"], 
+            dtype=self.dtype
         )
-        events = np.lib.recfunctions.unstructured_to_structured(events, self.dtype)
+        
         if "/dvs/imu" in topics.keys():
             imu = topics["/dvs/imu"]
             imu["ts"] = ((imu["ts"] - imu["ts"][0]) * 1e6).astype(int)