--- a/tonic/datasets/dsec.py
+++ b/tonic/datasets/dsec.py
@@ -9,6 +9,7 @@ from tonic.download_utils import (
     download_url,
     list_files,
 )
+from tonic.io import make_structured_array
 
 
 class DSEC(Dataset):
@@ -167,31 +168,23 @@ class DSEC(Dataset):
         events_left_file = h5py.File(
             os.path.join(base_folder, "events_left", "events.h5")
         )["events"]
-        events_left = np.column_stack(
-            (
+        events_left = make_structured_array(
                 events_left_file["x"][()],
                 events_left_file["y"][()],
                 events_left_file["t"][()],
                 events_left_file["p"][()],
-            )
-        )
-        events_left = np.lib.recfunctions.unstructured_to_structured(
-            events_left, self.dtype
+                dtype=self.dtype
         )
 
         events_right_file = h5py.File(
             os.path.join(base_folder, "events_right", "events.h5")
         )["events"]
-        events_right = np.column_stack(
-            (
+        events_right = make_structured_array(
                 events_right_file["x"][()],
                 events_right_file["y"][()],
                 events_right_file["t"][()],
                 events_right_file["p"][()],
-            )
-        )
-        events_right = np.lib.recfunctions.unstructured_to_structured(
-            events_right, self.dtype
+                dtype=self.dtype
         )
 
         # images