@biphasic 2022-04-28
@jrfinkbeiner @fabhertz95 many thanks for the benchmarking done. Good find! I replicated the three options with the following code
```python
import numpy as np
from numpy.lib import recfunctions

n_events = 100000
x = np.ones((n_events))
y = np.ones((n_events))
t = np.ones((n_events))
p = np.ones((n_events))

events_struct = np.dtype([("x", np.int16), ("y", np.int16), ("t", np.int64), ("p", bool)])

def make_structured_array(x, y, t, p):
    return np.fromiter(zip(x, y, t, p), dtype=events_struct)

def make_structured_array_proposed(x, y, t, p):
    struct_arr = np.empty_like(x, dtype=events_struct)
    struct_arr["x"] = x
    struct_arr["y"] = y
    struct_arr["t"] = t
    struct_arr["p"] = p
    return struct_arr

def make_structured_array_librec(x, y, t, p):
    return recfunctions.unstructured_to_structured(np.column_stack((x,y,t,p)), dtype=events_struct)
```

The result is as follows on my laptop:
![Screenshot from 2022-04-28 23-39-27](https://user-images.githubusercontent.com/11279568/165850204-9e485483-b371-4d68-8d7d-89e4eb40609c.png)


That's quite a big speed up although honestly I'm scratching my head about it a bit. In any case I prefer the fastest option available, which seems to be the proposed solution in this PR. For consistency it would be great to replace all occurences of `np.lib.recfunctions.unstructured_to_structured(np.column_stack(...))` with the new tonic.io.make_structured_array(x, y, t, p, dtype).

@jrfinkbeiner do you think you could do that modification?

Also the function `read_mnist_file` exists because NCALTECH101 and NMNIST use files encoded in a specific binary format.