@jrfinkbeiner 2022-05-03
@biphasic @fabhertz95 Sorry for this mess. Should I create a new pull request with a single commit? 
Didn't expect issues with the changes. 

In the meantime, I can already make you aware of the some changes:

I changed the function `make_structured_array` so that it takes a variable amount of arguments. (Forgot to update the docstring for now.) This is necessary for the audio datasets. Additionally, I am assuming now that the arguments are passed in the order that is present in the `dtype.names`. I chose the `*args` syntax (opposed to just having a list/tuple as first argument) as it is more similar to the precious syntax. The only issue I see is that now the dtype can not be given as positional argument, but it has to be a keyword argument. Therefore I included a check.

```python
def make_structured_array(*args, dtype=events_struct):
    assert not isinstance(args[-1], np.dtype), "The `dtype` must be provided as a keyword argument."
    names = dtype.names
    assert len(args) == len(names)
    struct_arr = np.empty_like(args[0], dtype=dtype)
    for arg,name in zip(args,names):
        struct_arr[name] = arg
    return struct_arr
```

Regarding the replacement of `np.lib.recfunctions.unstructured_to_structured` with `make_structured_array`: I only changed the instances where it was used in combination with a version of `np.stack`. If a full array already existed, I left it at with the old syntax (e.g. as in "dvsgestures.py"), because profiling indicated that this might be the faster option then. 

Let me know whether I should include other changes (apart from the ones to make the tests pass) and whether I should create a new "clean" pull request.