@jrfinkbeiner 2022-05-04
> Hi @jrfinkbeiner, I would add the changes in the test routines, so that the PR is clean.

@fabhertz95 sorry, I am not quite sure what this means. I am only aware of creating a new pull request based on a rebased commit to create a clean PR. And apparently rebase within the current PR also works? However, I've never done that. 

@biphasic @fabhertz95 From my side, I am happy now with the pull request. Let me know, what to do.