@fabhertz95 2022-05-04
> @fabhertz95 sorry, I am not quite sure what this means. I am only aware of creating a new pull request based on a rebased commit to create a clean PR. And apparently rebase within the current PR also works? However, I've never done that.

Don't worry, I did not explain myself well. I meant adding the modifications needed for the tests. Everything seems fine to me now :)