@biphasic 2022-05-05
nice, this one looks great, you made the code look cleaner. The make_structured_array function is now very general, I like that. Thanks a lot :+1: